strlcat fonctionnel !!!

+++ libft/ft_strcmp.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int		ft_strcmp(const char *s1, const char *s2)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] && s2[i] && ((unsigned char*)s1)[i] == ((unsigned char*)s2)[i])
